move reference binding to Env
