+++ packages/runtime/src/HandleScope.ts
@@ -1,4 +1,6 @@
+import type { Env } from './env'
 import type { Handle, HandleStore } from './Handle'
+import { External } from './Handle'
 
 /** @internal */
 export class HandleScope {
@@ -27,8 +29,11 @@ export class HandleScope {
     return h
   }
 
-  public addExternal (data: void_p): Handle<object> {
-    const h = this.handleStore.pushExternal(data)
+  public addExternal (envObject: Env, data: void_p): Handle<object> {
+    const value = new (External as any)()
+    const h = envObject.ctx.handleStore.push(value)
+    const binding = envObject.initObjectBinding(value)
+    binding.data = data
     this.end++
     return h
   }
