+++ packages/runtime/src/index.ts
@@ -6,11 +6,11 @@ export { CallbackInfo, CallbackInfoStack } from './CallbackInfo'
 export { createContext, Context } from './Context'
 export { Deferred, type IDeferrdValue } from './Deferred'
 export { DeferredStore } from './DeferredStore'
-export { type ILastError, Env } from './env'
+export { type ILastError, Env, type IReferenceBinding } from './env'
 export { EnvStore } from './EnvStore'
 export { EmnapiError, NotSupportWeakRefError, NotSupportBigIntError } from './errors'
 export { Finalizer } from './Finalizer'
-export { Handle, ConstHandle, HandleStore, type IReferenceBinding } from './Handle'
+export { Handle, ConstHandle, HandleStore } from './Handle'
 export { HandleScope } from './HandleScope'
 export { RefBase, Ownership } from './RefBase'
 export { Persistent } from './Persistent'
